Move sublime.Kind tuple definitions into constants module
